support --read-only containers
This allows containers to be run like

    docker run -dit \
        --name postgresql-1-ctl-1 \
        -v postgresql-1-ctl-1:/var/lib/postgres-xl \
        -v postgresql-1-ssh:/var/lib/postgres-xl/.ssh \
        --tmpfs /tmp \
        --read-only \
        postgres-xl

For non-`ctl` nodes, `/var/lib/postgres-xl/.ssh` volume can also be
read-only.

Note that there is a `/etc/ssh` volume, which should be maintained for
the lifetime of that node, even if the service is restarted (since it
contains host SSH keys).