[#12] proxy: remove from default config
Remove `proxy_1`, `proxy_2` from default config, instead connecting
directly to `gtm_1`. This simplifies the setup somewhat, and paves the
way for future healthchecks work (I don't have healthchecks I'm happy
with for proxy). Removing as part of this round of work, so I don't
spend time adding default config for something I'm about to remove. :)
To clarify, I intend to continue supporting Proxy, but not to use it
for examples or the default config.