build: add docker-compose.image.yml
I think some people thing that the way to use Postgres-XL Docker is to
build it themselves from the Compose file. This is not the case; the
images are published to Docker Hub, and those should normally be used
instead. There's no need to compile this locally, unless you actually
want to develop Postgres-XL Docker (or possibly Postgres-XL) itself.