[#25] build: rework healthchecks, adding bootstrapping phase
Docker Swarm usage complicates healthchecks, since the containers are
unroutable until healthchecks pass, but for healthchecks to pass, the
cluster has to be not only routing correctly, but initialised.

The Swarm example init script, `bin/init-eg-swarm`, overcame this by
overriding the healthchecks at the service level, allowing them to pass.
Then, the cluster was expected to be redeployed, in order for the
healthchecks to pass. This, however, is not only rather unpleasant, but
still doesn't solve issues in all cases, or indeed when using other
schedulers or approaches.

Rework healthchecks, adding a bootstrapping phase. Now, when containers
come up, healthchecks will pass trivially, even if the cluster hasn't
been initted. However, on every Coord healthcheck, an attempt will be
made to create the user, database, and table necessary for the proper
healthcheck to pass. If this fails, however, it won't cause the
healthcheck to fail. Then, whenever the healthcheck passes for the first
time properly, necessarily after init, the bootstrapping phase is
complete, and all future Coord and Data healthchecks will check
properly.

This also has the side-effect of optimising the healthcheck to be
lighter, and allows also for the `bin/init-eg-swarm` script to avoid the
unpleasant service healthcheck overrides. This also means that a
redeploy is no longer needed.

Also optimise `bin/init-eg` (non-Swarm) example init script, to not
restart the containers, since this is not necessary.