doc: update LICENSE, README to note other contributors
Sorry, but I don't think naming every contributor is maintainable --
only the maintainers, or at least those with a commit-bit. But many
thanks for your contributions, and this is reflected not only by the PR
records, but also by these explicit mentions in the README and LICENSE
linking to a page which lists your individual handle automagically. :)